Update issues review comments
