+++ frontend/.env.example
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_API_URL="YOUR_API_URL "
+NEXT_PUBLIC_API_URL="YOUR_API_URL"
