+++ frontend/package.json
@@ -26,7 +26,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "eslint": "^8",
+    "eslint": "^9.16.0",
     "eslint-config-next": "15.0.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
