+++ frontend/src/app/layout.tsx
@@ -10,7 +10,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
+    <html lang="ja">
       <body>
         <Provider store={store}>{children}</Provider>
       </body>
